Add files via upload
In this project, I've written a program in C language,  which gets the extra worked hours from the user as well as the additional payment that employee will receive. If the extra hours of work exceeds 8 hours, the additional payment and the extra work hours inputted will be multiplied and added to their initial salary and will be outputted.